Big changes, PHP navbar and footer. Changed JS
Went ahead and changed all pages, that needed to be, to PHP and made the
navbar and footer just an includes to get rid of redundant code. Navbar
now uses $_GET variables to determine which page it is on and set the
'active' state accordingly. The javascript changes are not complete yet
only thing I have changed is the way the modal works for the first bike
in mtb.php.